Arrondir les résultats d'un tableau

tab_round() permet d'arrondir les données des cellules, et éventuellement de gérer la cohérence des sommes à 100 % pour les pourcentages lignes et colonnes, en proposant différentes règles de décision.

tab_round(
  tab,
  n = 0,
  guarantee_100 = FALSE,
  option = "lowest_lie",
  verbose = FALSE
)

Arguments

tab

tableau croisé issu de tab_build

n

nombre de décimales à conserver - par défaut : 0 (arrondi à l'entier) Ce paramètre peut prendre des valeurs positives, ou négatives pour arrondir à la dizaine, la centaine, etc :

0, 1, 2, ... : entier, 1 décimale, 2 décimales, etc...

-1 : arrondi à la dizaine

-2 : arrondi à la centaine

"etc..." : ...

guarantee_100

pour les arrondis de pourcentages lignes ou colonnes, faut-il gérer la cohérence des arrondis au niveau n ? - par défaut : FALSE

option

option à retenir pour le calcul de la garantie à 100 - par défaut : "lowest_lie" Ce paramètre peut prendre les valeurs suivantes :

"lowest_lie" : ...

"closest" : ...

verbose

affichage ou non des détails sur ce qui est fait pour guarantee_100 - par défaut : FALSE

Value

tableau avec cellules arrondies

Examples

library(dplyr)

prenoms |>
  filter(annais == "2020",
         lib_reg %in% c("Occitanie", "Pays de la Loire")) |>
  tab_build(var_rows = c(lib_reg, sexe),
            var_cols = prenom,
            var_stat = nombre,
            lab_total = "Total",
            stat = "sum") |>
  tab_pct(pct = "row") |>
  tab_round(n = 0,
            guarantee_100 = TRUE,
            option = "closest",
            verbose = TRUE)
#> guarantee_100 : la mise en cohérence des marges en colonnes reste à effectuer par l'utilisateur (ok en lignes)
#> >> La somme fait déjà 100 avec le bon niveau d'arrondi.
#> >> La somme fait 99
#> >> On change le 8.4 en position 5 par 9
#> >> La somme fait 101
#> >> On change le 17.57 en position 4 par 17
#> >> La somme fait 99
#> >> On change le 39.47 en position 4 par 40
#> >> La somme fait déjà 100 avec le bon niveau d'arrondi.
#> >> La somme fait déjà 100 avec le bon niveau d'arrondi.
#> >> La somme fait 101
#> >> On change le 5.53 en position 7 par 5
#> # A tibble: 7 × 10
#>   lib_reg          sexe   Alix Camille Charlie  Eden Louison   Noa Sasha Total
#> * <chr>            <chr> <dbl>   <dbl>   <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl>
#> 1 Pays de la Loire G         3      18      15    31      16    11     6   100
#> 2 Pays de la Loire F        26      31      26     4       9     1     3   100
#> 3 Pays de la Loire Total    14      25      21    17      12     6     5   100
#> 4 Occitanie        G         0      17      13    40       1    19    10   100
#> 5 Occitanie        F        26      27      34     5       2     3     3   100
#> 6 Occitanie        Total    13      22      24    22       2    11     6   100
#> 7 Total            Total    14      24      22    20       7     8     5   100